chore(code-preview): remove obsolete handwritten generic lexers
