+++ src/preview/code/backends/mod.rs
@@ -1,4 +1,3 @@
 pub(super) mod custom;
-pub(super) mod legacy;
 pub(super) mod plain;
 pub(super) mod syntect;
