+++ src/preview/mod.rs
@@ -4,7 +4,6 @@ mod container;
 mod directory;
 mod dispatch;
 mod document;
-mod highlighting;
 mod markdown;
 mod structured;
 mod text;
